Pass header provider values via context; Header just renders them
Dialog builds the values array ([HeadingContext, ...], [ButtonContext, ...]) and
hands it to Header through HeaderContext; Header renders a Provider around its own
children with it. Typed explicitly as correlated [Context<T>, T] pairs so it
passes through context into Provider with no cast and no any. content.tsx no
longer imports the Close icon, locale or React Aria's HeadingContext.
